+++ app/src/main/java/com/ruanyun/australianews/model/VipDetailIfo.java
@@ -76,8 +76,8 @@ public class VipDetailIfo {
         private int contentType;
         private String createTimeStr;
         private int uv;
-        private double specialOffercny;
-        private double specialOfferaud;
+        private float specialOffercny;
+        private float specialOfferaud;
         private String afnNewsType;
         private String commentInfoPagination;
         private int isDelete;
@@ -86,9 +86,9 @@ public class VipDetailIfo {
         private String updateTime;
         private int commentCount;
         private String keyWord;
-        private double normalPriceaud;
+        private float normalPriceaud;
         private String beginDate;
-        private double normalPricecny;
+        private float normalPricecny;
         private String createTime;
         private int isTop;
         private int isHot;
@@ -208,11 +208,11 @@ public class VipDetailIfo {
             this.uv = uv;
         }
 
-        public void setSpecialOffercny(double specialOffercny) {
+        public void setSpecialOffercny(float specialOffercny) {
             this.specialOffercny = specialOffercny;
         }
 
-        public void setSpecialOfferaud(double specialOfferaud) {
+        public void setSpecialOfferaud(float specialOfferaud) {
             this.specialOfferaud = specialOfferaud;
         }
 
@@ -248,7 +248,7 @@ public class VipDetailIfo {
             this.keyWord = keyWord;
         }
 
-        public void setNormalPriceaud(double normalPriceaud) {
+        public void setNormalPriceaud(float normalPriceaud) {
             this.normalPriceaud = normalPriceaud;
         }
 
@@ -256,7 +256,7 @@ public class VipDetailIfo {
             this.beginDate = beginDate;
         }
 
-        public void setNormalPricecny(double normalPricecny) {
+        public void setNormalPricecny(float normalPricecny) {
             this.normalPricecny = normalPricecny;
         }
 
@@ -384,11 +384,11 @@ public class VipDetailIfo {
             return uv;
         }
 
-        public double getSpecialOffercny() {
+        public float getSpecialOffercny() {
             return specialOffercny;
         }
 
-        public double getSpecialOfferaud() {
+        public float getSpecialOfferaud() {
             return specialOfferaud;
         }
 
@@ -424,7 +424,7 @@ public class VipDetailIfo {
             return keyWord;
         }
 
-        public double getNormalPriceaud() {
+        public float getNormalPriceaud() {
             return normalPriceaud;
         }
 
@@ -432,7 +432,7 @@ public class VipDetailIfo {
             return beginDate;
         }
 
-        public double getNormalPricecny() {
+        public float getNormalPricecny() {
             return normalPricecny;
         }
